requirements.txt
langchain
langchain-groq
langchain-chroma
langchain-community
langchain-text-splitters
langgraph
langgraph-checkpoint-sqlite
langsmith

huggingface_hub
chromadb

streamlit
requests
pypdf
python-dotenv
pytest
ruff
